Tests: refactor: extract `ExecContext` into the separate module.
